requirements.txt
websockets>=11.0.3
python-dotenv>=1.0.0
mcp>=1.8.1
pydantic>=2.11.4
mcp-proxy>=0.8.2
pyatv>=0.14.0
